requirements.txt
aiohttp
aiosqlite
discord.py
python-dotenv
pyautogen